+++ src/MainTransformer/MainTransformerInterface.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\MainTransformer;
 
+use Rekalogika\Mapper\Context\Context;
 use Symfony\Component\PropertyInfo\Type;
 
 interface MainTransformerInterface