+++ src/Mapper.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper;
 
 use Rekalogika\Mapper\Exception\UnexpectedValueException;
-use Rekalogika\Mapper\MainTransformer\Context;
+use Rekalogika\Mapper\Context\Context;
 use Rekalogika\Mapper\MainTransformer\MainTransformerInterface;
 use Rekalogika\Mapper\Util\TypeFactory;