+++ src/Transformer/Contracts/TransformerInterface.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Transformer\Contracts;
 
-use Rekalogika\Mapper\MainTransformer\Context;
+use Rekalogika\Mapper\Context\Context;
 use Symfony\Component\PropertyInfo\Type;
 
 interface TransformerInterface