+++ tests/Fixtures/MethodMapper/MoneyDto.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper\Tests\Fixtures\MethodMapper;
 
 use Brick\Money\Money;
-use Rekalogika\Mapper\MainTransformer\Context;
+use Rekalogika\Mapper\Context\Context;
 use Rekalogika\Mapper\MethodMapper\MapFromObjectInterface;
 use Rekalogika\Mapper\MethodMapper\MapToObjectInterface;
 use Rekalogika\Mapper\MethodMapper\SubMapperInterface;