+++ tests/Fixtures/Transformer/MoneyToMoneyDtoTransformer.php
@@ -15,7 +15,7 @@ namespace Rekalogika\Mapper\Tests\Fixtures\Transformer;
 
 use Brick\Money\Money;
 use Rekalogika\Mapper\Exception\InvalidArgumentException;
-use Rekalogika\Mapper\MainTransformer\Context;
+use Rekalogika\Mapper\Context\Context;
 use Rekalogika\Mapper\Tests\Fixtures\Money\MoneyDto;
 use Rekalogika\Mapper\Transformer\Contracts\TransformerInterface;
 use Rekalogika\Mapper\Transformer\Contracts\TypeMapping;